+++ android/src/oldarch/java/com/reactnativestripesdk/NativeStripeSdkModuleSpec.java
@@ -278,9 +278,25 @@ public abstract class NativeStripeSdkModuleSpec extends ReactContextBaseJavaModu
 
   @ReactMethod
   @DoNotStrip
+  public abstract void authWebViewDeepLinkHandled(String id, Promise promise);
+
+  @ReactMethod
+  @DoNotStrip
+  public abstract void storeStripeConnectDeepLink(String url, Promise promise);
+
+  @ReactMethod
+  @DoNotStrip
+  public abstract void pollAndClearPendingStripeConnectUrls(Promise promise);
+
+  @ReactMethod
+  @DoNotStrip
   public abstract void addListener(String eventType);
 
   @ReactMethod
   @DoNotStrip
   public abstract void removeListeners(double count);
+
+  @ReactMethod
+  @DoNotStrip
+  public abstract void createRadarSession(Promise promise);
 }